change all float price with BigInteger

+++ database/migrations/2022_01_22_212716_create_invoices_table.php
@@ -24,9 +24,9 @@ class CreateInvoicesTable extends Migration
             $table->string('bl_code')->nullable();
             $table->string('bc_code')->nullable();
 
-            $table->float('price_ht')->default(0);
-            $table->float('price_total')->default(0);
-            $table->float('price_tva')->default(0);
+            $table->unsignedBigInteger('price_ht')->default(0);
+            $table->unsignedBigInteger('price_total')->default(0);
+            $table->unsignedBigInteger('price_tva')->default(0);
 
             $table->integer('status')->default(Response::INVOICE_EN_ATTENTE);
